+++ src/Javinizer/Private/Get-VideoFile.ps1
@@ -3,18 +3,20 @@ function Get-VideoFile {
     param (
         [Parameter(Position = 0)]
         [string]$Path,
-        [int]$FileSize = 10,
-        [switch]$Recurse
+        [int]$FileSize,
+        [switch]$Recurse,
+        [object]$Settings
     )
 
     begin {
         Write-Debug "[$($MyInvocation.MyCommand.Name)] Function started"
+        $FileSize = $Settings.General.'minimum-filesize-to-sort'
     }
 
     process {
         # Test if Path is a directory or item
-        if (Test-Path -Path (Get-Item $Path) -PathType Container) {
-            $files = Get-ChildItem -Path $Path -Recurse:$Recurse | Where-Object {
+        if (Test-Path -Path (Get-Item -LiteralPath $Path) -PathType Container) {
+            $files = Get-ChildItem -LiteralPath $Path -Recurse:$Recurse | Where-Object {
                 $_.Name -like '*.mp4'`
                     -or $_.Name -like '*.avi'`
                     -or $_.Name -like '*.mkv'`
@@ -23,15 +25,12 @@ function Get-VideoFile {
                     -or $_.Name -like '*.mov'`
                     -or $_.Name -like '*.m4v'`
                     -or $_.Name -like '*.rmvb'`
-                    -and $_.Name -notlike '*1pon*'`
-                    -and $_.Name -notlike '*carib*' `
-                    -and $_.Name -notlike '*fc2*'`
-                    -and $_.Name -notlike '*trailer'
-                #-and $_.Length -ge ($FileSize * 1MB)`
+                    -and $_.Name -notlike '*trailer'`
+                    -and $_.Length -ge ($FileSize * 1MB)
             }
             # Test if the path is a file
-        } elseif (Test-Path -Path (Get-Item $Path) -PathType Leaf) {
-            $files = Get-Item -Path $Path | Where-Object {
+        } elseif (Test-Path -Path (Get-Item -LiteralPath $Path) -PathType Leaf) {
+            $files = Get-Item -LiteralPath $Path | Where-Object {
                 $_.Name -like '*.mp4'`
                     -or $_.Name -like '*.avi'`
                     -or $_.Name -like '*.mkv'`
@@ -40,14 +39,11 @@ function Get-VideoFile {
                     -or $_.Name -like '*.mov'`
                     -or $_.Name -like '*.m4v'`
                     -or $_.Name -like '*.rmvb'`
-                    -and $_.Name -notlike '*1pon*'`
-                    -and $_.Name -notlike '*carib*' `
-                    -and $_.Name -notlike '*fc2*'`
-                    -and $_.Name -notlike '*trailer'
-                #-and $_.Length -ge ($FileSize * 1MB)`
+                    -and $_.Name -notlike '*trailer'`
+                    -and $_.Length -ge ($FileSize * 1MB)
             }
         } else {
-            throw "The path specified is neither directory nor file"
+            throw "[$($MyInvocation.MyCommand.Name)] The path specified is neither directory nor file"
         }
 
         Write-Output $files